requirements.txt
fastapi
uvicorn
starlette

pyaudio
webrtcvad-wheels
numpy
torch
ctranslate2
faster_whisper

nltk
langchain
langchain-groq
langchain-community
python-dotenv
langchain-core

git+https://github.com/huggingface/parler-tts.git

soundfile 
sounddevice 

protobuf
